Started adding sweepstakes winner selection
